Moved the circles introduction into an 'introduction' containing folder.
