+++ assets/pages/regular-contributions/components/regularContributionsThankYouPage.jsx
@@ -8,7 +8,7 @@ import { connect } from 'react-redux';
 import SimpleHeader from 'components/headers/simpleHeader/simpleHeader';
 import PageSection from 'components/pageSection/pageSection';
 import Footer from 'components/footer/footer';
-import CirclesIntroduction from 'components/circlesIntroduction/circlesIntroduction';
+import CirclesIntroduction from 'components/introduction/circlesIntroduction';
 import QuestionsContact from 'components/questionsContact/questionsContact';
 import SpreadTheWord from 'components/spreadTheWord/spreadTheWord';
 import DirectDebitGuarantee from 'components/directDebit/directDebitForm/directDebitGuarantee';
