+++ assets/stylesheets/main.scss
@@ -14,7 +14,7 @@
 @import '../components/directDebit/directDebitPopUpForm/directDebitPopUpForm';
 @import '../components/directDebit/directDebitForm/directDebitForm';
 @import '../components/checkboxInput/checkboxInput';
-@import '../components/circlesIntroduction/circlesIntroduction';
+@import '../components/introduction/circlesIntroduction';
 @import '../components/contribute/contribute';
 @import '../components/contributionPaymentCtas/contributionPaymentCtas';
 @import '../containerisableComponents/contributionSelection/contributionSelection';
